requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Google Cloud
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0
google-cloud-bigquery==3.26.0
google-cloud-resource-manager==1.13.1
google-cloud-iam==2.15.2
firebase-admin==6.5.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
tenacity==9.0.0
